requirements.txt
tornado==4.3
msgpack-python==0.4.6
pymongo==2.8
python-dateutil==2.4.2
motor==0.4.1
nose==1.3.7
